+++ crates/adapters/src/format/avro/output.rs
@@ -87,6 +87,8 @@ struct AvroEncoder {
     buffer: Vec<u8>,
     /// Count of skipped deletes, used to rate-limit error messages.
     skipped_deletes: usize,
+    /// `True` if the serialized result should not include the schema ID.
+    skip_schema_id: bool,
 }
 
 impl AvroEncoder {
@@ -219,7 +221,9 @@ impl AvroEncoder {
         };
 
         let mut buffer = vec![0u8; 5];
-        buffer[1..].clone_from_slice(&schema_id.to_be_bytes());
+        if !config.skip_schema_id {
+            buffer[1..].clone_from_slice(&schema_id.to_be_bytes());
+        }
 
         Ok(Self {
             endpoint_name: endpoint_name.to_string(),
@@ -227,6 +231,7 @@ impl AvroEncoder {
             schema,
             buffer,
             skipped_deletes: 0,
+            skip_schema_id: config.skip_schema_id,
         })
     }
 }
@@ -276,8 +281,13 @@ impl Encoder for AvroEncoder {
                 let mut avro_buffer = to_avro_datum(&self.schema, avro_value)
                     .map_err(|e| anyhow!("error serializing Avro value: {e}"))?;
 
-                // 5 is the length of the Avro message header (magic byte + 4-byte schema id).
-                self.buffer.truncate(5);
+                if !self.skip_schema_id {
+                    // 5 is the length of the Avro message header (magic byte + 4-byte schema id).
+                    self.buffer.truncate(5);
+                } else {
+                    self.buffer.clear();
+                }
+
                 self.buffer.append(&mut avro_buffer);
                 self.output_consumer.push_buffer(&self.buffer, 1);